% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_vault.R
\name{explore_vault}
\alias{explore_vault}
\title{Explore a vault}
\usage{
explore_vault(con = NULL)
}
\arguments{
\item{con}{A \code{vault_pipe} object created by \code{open_vault()}.}
}
\value{
A \code{character} vector of table names present in the VegVault database.
}
\description{
Get the names of all tables in the connected VegVault database.
This function can be called at any stage of the pipeline.
}
