% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract data from a plan}
\usage{
extract_data(con, return_raw_data = FALSE, verbose = TRUE)
}
\arguments{
\item{con}{A vault connection}

\item{return_raw_data}{A logical indicating whether to return raw data or
without any processing. Default is \code{FALSE}.}

\item{verbose}{A logical indicating whether output additional information.
Default is \code{TRUE}.}
}
\value{
A data.frame
}
\description{
Extract data from a vault connection. The data is first
processed with columns translated into more redable variants.
The data is then packed into a single tibble with nested data.frames,
containing information about:
\itemize{
\item \code{data_samples} for sample metadata
\item \code{data_samples_community} for community (vegetation) data
\item \code{data_traits} for trait data
\item \code{data_abiotic} for abiotic data
}
}
