% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abiotic.R
\name{get_abiotic}
\alias{get_abiotic}
\title{Get abiotic data}
\usage{
get_abiotic(con = NULL)
}
\arguments{
\item{con}{A connection object created by \code{open_vault()}}
}
\value{
A connection object with abiotic data
}
\description{
Get abiotic data from the Vault database
}
