% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abiotic_data.R
\name{get_abiotic_data}
\alias{get_abiotic_data}
\title{Get abiotic data}
\usage{
get_abiotic_data(
  con = NULL,
  mode = c("nearest", "mean", "median"),
  limit_by_distance_km = 50,
  limit_by_age_years = 5000,
  verbose = TRUE
)
}
\arguments{
\item{con}{A connection object created by \code{open_vault()}}
}
\value{
A connection object with abiotic data
}
\description{
Get abiotic data from the Vault database
}
