% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_uncertainty.R
\name{get_age_uncertainty}
\alias{get_age_uncertainty}
\title{Get age uncertainties for samples}
\usage{
get_age_uncertainty(con = NULL)
}
\arguments{
\item{con}{A \code{vault_pipe} object created by \code{open_vault()}. Must already contain
\code{sample_id} in the data, i.e. \code{get_samples()} must have been called
earlier in the pipe.}
}
\value{
A \code{tibble} with the columns \code{sample_id}, \code{iteration}, and
\code{age_uncertainty}, restricted to the samples present in the
incoming pipe.
}
\description{
For each Sample, get age uncertainty estimates from the Vault database.
Age uncertainties represent multiple iterations of age-depth modelling
and are stored in the \code{SampleUncertainty} table.
}
\seealso{
\code{\link[=open_vault]{open_vault()}}, \code{\link[=get_samples]{get_samples()}}
}
