% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_common_dataset_type.R
\name{get_common_dataset_type}
\alias{get_common_dataset_type}
\title{Get common dataset type}
\usage{
get_common_dataset_type(data_source, vec_dataset_type, verbose = TRUE)
}
\arguments{
\item{data_source}{A tbl object containing the dataset}

\item{vec_dataset_type}{A character vector of dataset types to filter}

\item{verbose}{A logical value to print messages}
}
\value{
A character vector of common dataset type
}
\description{
Get the common dataset type between the data and
the selected dataset type
}
\keyword{internal}
