% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{get Datasets}
\usage{
get_datasets(con)
}
\arguments{
\item{con}{A connection object created by \code{open_vault()}}
}
\value{
A \code{vault_pipe} object with the Datasets table
}
\description{
Get the Datasets table from the Vault database
}
