% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_references.R
\name{get_references}
\alias{get_references}
\title{Get references for selected VegVault Plan}
\usage{
get_references(
  con = NULL,
  path = NULL,
  type = c("Dataset", "DatasetSource", "DatasetSourceType", "SamplingMethod", "Sample",
    "Taxon", "Trait", "AbioticVariable")
)
}
\arguments{
\item{con}{A VegVault connection (a plan).}

\item{type}{A character vector specifying the type of references to be
extracted. The following types are available: \code{Dataset}, \code{DatasetSource},
\code{DatasetSourceType}, \code{SamplingMethod}, \code{Sample}, \code{Taxon}, \code{Trait}, and
\code{AbioticVariable}.}
}
\value{
A tibble containing all references for the selected data compilation.
}
\description{
This function extracts all references for selected
VegVault extraction plan (i,e., this function should be used before
extracting data from the plan).
}
