% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples.R
\name{get_samples}
\alias{get_samples}
\title{Get samples}
\usage{
get_samples(con = NULL)
}
\arguments{
\item{con}{A connection to the Vault database}
}
\value{
A connection to the Vault database
}
\description{
Get samples from the Vault database
}
