% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_vault.R
\name{open_vault}
\alias{open_vault}
\title{Open a vault}
\usage{
open_vault(path)
}
\arguments{
\item{path}{A character string of the path to the SQLite database}
}
\value{
A list with two elements: \code{data} and \code{db_con}
}
\description{
Open a connection to a SQLite database
}
