% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_data.R
\name{pack_data}
\alias{pack_data}
\title{Pack the data into nested tibbles}
\usage{
pack_data(sel_data = NULL, verbose = TRUE)
}
\arguments{
\item{sel_data}{A tibble with data}

\item{verbose}{A logical indicating whether to output additional information.
Default is \code{TRUE}.}
}
\value{
A tibble with nested data.frames
}
\description{
Pack the data into nested tibbles for easier access and
manipulation. Only specific columns are selected and packed into the
following nested tibbles:
\itemize{
\item \code{data_samples} for sample metadata
\item \code{data_community} for community (vegetation) data
\item \code{data_traits} for trait data
\item \code{data_abiotic} for abiotic data
When the \code{sel_data} does not contain the required columns, the function
will skip the respective data type.
}
}
\keyword{internal}
