% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_abiotic_var_by_id.R
\name{select_abiotic_var_by_id}
\alias{select_abiotic_var_by_id}
\title{Select abiotic variables by id}
\usage{
select_abiotic_var_by_id(con = NULL, sel_var_id = NULL)
}
\arguments{
\item{con}{A class of \code{vault_pipe}}

\item{sel_var_id}{A character vector of abiotic variable ids}
}
\value{
A class of \code{vault_pipe}
}
\description{
Filter the data by abiotic variables id
}
