% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_abiotic_var_by_name.R
\name{select_abiotic_var_by_name}
\alias{select_abiotic_var_by_name}
\title{Select abiotic variables by name}
\usage{
select_abiotic_var_by_name(con = NULL, sel_var_name = NULL)
}
\arguments{
\item{con}{A class of \code{vault_pipe}}

\item{sel_var_name}{A character vector of abiotic variable names}
}
\value{
A class of \code{vault_pipe}
}
\description{
Filter the data by abiotic variables name
}
