% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_dataset_by_geo.R
\name{select_dataset_by_geo}
\alias{select_dataset_by_geo}
\title{Select dataset by geographical location}
\usage{
select_dataset_by_geo(
  con = NULL,
  long_lim = c(-180, 180),
  lat_lim = c(-90, 90),
  sel_dataset_type = c("vegetation_plot", "fossil_pollen_archive", "gridpoints"),
  verbose = TRUE
)
}
\arguments{
\item{con}{A class of \code{vault_pipe}}

\item{long_lim}{A numeric vector of length 2 specifying the longitude limits}

\item{lat_lim}{A numeric vector of length 2 specifying the latitude limits}

\item{sel_dataset_type}{(optional) A character vector specifying the dataset types to filter}

\item{verbose}{A logical value specifying whether to print messages}
}
\value{
A class of \code{vault_pipe}
}
\description{
Filter the dataset by geographical location
}
\details{
There is an option to filter only specific dataset types (as some dataset types may not have lat/long information).
To do this, use the \code{sel_dataset_type} argument.
}
