% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_dataset_by_type.R
\name{select_dataset_by_type}
\alias{select_dataset_by_type}
\title{get Datasets by selected type}
\usage{
select_dataset_by_type(
  con,
  sel_dataset_type = c("vegetation_plot", "fossil_pollen_archive", "traits",
    "gridpoints")
)
}
\arguments{
\item{con}{A connection object created by \code{open_vault()}}

\item{sel_dataset_type}{A character vector of dataset types to filter}
}
\value{
A \code{vault_pipe} object with the filtered Datasets table
}
\description{
Filter the Datasets by a selected DatasetType
}
