% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_samples_by_age.R
\name{select_samples_by_age}
\alias{select_samples_by_age}
\title{Select samples by age}
\usage{
select_samples_by_age(
  con = NULL,
  age_lim = c(-Inf, Inf),
  sel_dataset_type = c("vegetation_plot", "fossil_pollen_archive", "gridpoints"),
  verbose = TRUE
)
}
\arguments{
\item{con}{A connection object created by \code{open_vault()}}

\item{age_lim}{A numeric vector of length 2.
The minimum and maximum age of the samples to be selected.}

\item{sel_dataset_type}{(optional) A character vector specifying the dataset types to filter}

\item{verbose}{A logical value specifying whether to print messages}
}
\value{
A class of \code{vault_pipe}
}
\description{
Select samples by age criteria
}
\details{
There is an option to filter only specific dataset types (as some dataset types may not have age information).
To do this, use the \code{sel_dataset_type} argument.
}
