% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_taxa_by_name.R
\name{select_taxa_by_name}
\alias{select_taxa_by_name}
\title{Only select taxa by name}
\usage{
select_taxa_by_name(con = NULL, sel_taxa = NULL)
}
\arguments{
\item{con}{A connection to the Vault database}

\item{sel_taxa}{A character vector of taxa names}
}
\value{
A vault_pipe object
}
\description{
Filter the Samples in dataset by taxa name
}
