% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_traits_by_domain_name.R
\name{select_traits_by_domain_name}
\alias{select_traits_by_domain_name}
\title{Select traits by domain name}
\usage{
select_traits_by_domain_name(con = NULL, sel_domain = NULL)
}
\arguments{
\item{con}{A connection object created by \code{open_vault()}.}

\item{sel_domain}{A character vector of the selected Trait Domain name.}
}
\value{
A list of class `vault_pipe
}
\description{
Filters the data for the selected Trait Domain name.
}
