% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_taxa.R
\name{classify_taxa}
\alias{classify_taxa}
\title{Classify taxa}
\usage{
classify_taxa(
  data_source = NULL,
  sel_con = NULL,
  to = c("original", "species", "genus", "family")
)
}
\arguments{
\item{data_source}{A SQL table that contains \code{taxon_id} column.}

\item{sel_con}{A connection to the Vault database.}

\item{to}{A character vector that specifies the classification level to
classify the taxa. The options are \code{original}, \code{species}, \code{genus}, and
\code{family}.}
}
\value{
A SQL table with the classified taxa.
}
\description{
Classify taxa in a SQL table to a specific level based on the
classification table in the Vault database.
}
\keyword{internal}
