% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract data from a plan}
\usage{
extract_data(con, return_raw_data = FALSE, ...)
}
\arguments{
\item{con}{A vault connection}

\item{return_raw_data}{A logical indicating whether to return raw data or
without any processing. Default is \code{FALSE}.}

\item{...}{Additional argument passed to \code{pack_data()}}
}
\value{
A data.frame
}
\description{
Extract data from a vault connection
}
