% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract data from a plan}
\usage{
extract_data(con)
}
\arguments{
\item{con}{A vault connection}
}
\value{
A data.frame
}
\description{
Extract data from a vault connection
}
