% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_references.R
\name{get_references}
\alias{get_references}
\title{Get references for selected VegVault Plan}
\usage{
get_references(
  con = NULL,
  type = c("Dataset", "DatasetSource", "DatasetSourceType", "SamplingMethod", "Sample",
    "Taxon", "Trait", "AbioticVariable"),
  get_source = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{con}{A VegVault connection (a plan).}

\item{type}{A character vector specifying the type of references to be
extracted. The function will try to extract all possible reference types.
The following types are available: \code{Dataset}, \code{DatasetSource},
\code{DatasetSourceType}, \code{SamplingMethod}, \code{Sample}, \code{Taxon}, \code{Trait}, and
\code{AbioticVariable}.}

\item{get_source}{A logical indicating whether to extract the source of each
references}

\item{verbose}{A logical indicating whether to print messages about the
progress of the function. Default is \code{TRUE}.}
}
\value{
A tibble containing all references for the selected data compilation.
}
\description{
This function extracts all references for selected
VegVault extraction plan (i,e., this function should be used before
extracting data from the plan).
}
