% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_references.R
\name{get_references}
\alias{get_references}
\title{Get references for selected data compilation}
\usage{
get_references(
  data_source = NULL,
  path = NULL,
  type = c("Dataset", "DatasetSource", "DatasetSourceType", "SamplingMethod", "Sample",
    "Taxon", "Trait", "AbioticVariable"),
  verbose = TRUE
)
}
\arguments{
\item{data_source}{A tibble containing the data compilation, which
has been obtained using the \code{extract_data()} function.}

\item{path}{A character string specifying the path to the VegVault file.}

\item{type}{A character vector specifying the type of references to be
extracted. The function will try to extract all possible reference types.
The following types are available: \code{Dataset}, \code{DatasetSource},
\code{DatasetSourceType}, \code{SamplingMethod}, \code{Sample}, \code{Taxon}, \code{Trait}, and
\code{AbioticVariable}.}

\item{verbose}{A logical indicating whether to print messages about the
progress of the function. Default is \code{TRUE}.}
}
\value{
A tibble containing all references for the selected data compilation.
}
\description{
This function extracts all references for selected
data compilation, which have been extracted from VegVault.
}
