% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa.R
\name{get_taxa}
\alias{get_taxa}
\title{Get taxa information}
\usage{
get_taxa(con = NULL, classify_to = c("original", "species", "genus", "family"))
}
\arguments{
\item{con}{A connection to the Vault database}

\item{classify_to}{A character vector specifying the taxonomic level to classify
the taxa to. The options are \code{original}, \code{species}, \code{genus}, and \code{family}.}
}
\value{
A \code{vault_pipe} object with the data and the connection to the Vault database
}
\description{
For each Sample, get information about Taxa and their abundances
from the Vault database
}
