% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_traits.R
\name{get_traits}
\alias{get_traits}
\title{Get trait values for the data}
\usage{
get_traits(
  con = NULL,
  classify_to = c("original", "species", "genus", "family"),
  verbose = TRUE
)
}
\arguments{
\item{con}{A connection object created using \code{open_vault()}}

\item{classify_to}{A character vector specifying the taxonomic level
to classify the taxa to. Default is \code{original}}

\item{verbose}{A logical value indicating whether to print messages.
Default is \code{TRUE}}
}
\value{
A connection object with the data and database connection
}
\description{
Get trait values for the data from the Vault database
}
\details{
If the function is used after \code{get_taxa()}, the trait values will be
returned only for the taxa present in the data. If you prefer to
return all trait values, we recommend using \code{get_traits()} before
\code{get_taxa()} in the pipe.
In addition, it is important to set \code{classify_to} to the same value as
in \code{get_taxa()}.
}
