% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaultkeepr-package.R
\docType{package}
\name{vaultkeepr-package}
\alias{vaultkeepr}
\alias{vaultkeepr-package}
\title{vaultkeepr: Providing the Interface to Access the VegVault Database}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ondrej Mottl \email{ondrej.mottl@gmail.com} (\href{https://orcid.org/0000-0002-9796-5081}{ORCID})

}
\keyword{internal}
